summarizing simulation results

+++ condor/worker 1.sh
@@ -1,6 +1,12 @@
-mkdir output
-#mkdir logs
-Rscript code/simulations.r $*
+if [ -f /usr/bin/R ]
+then
+	mkdir output
+	
+	Rscript code/simulations.r $*
 
-tar -zcvf output/output$1-$2.tgz2 output/*.csv
-rm output/*.csv
+	tar -zcvf output/output$1-$2.tgz2 output/*.csv
+	rm output/*.csv
+	exit 0
+else 
+         exit 1
+fi
